110307 Upgrade installation config scripts for Java 11 Override existing host for prevent Tomcat error "Child name '...' is not unique"

+++ install-web.sh
@@ -35,6 +35,14 @@ fi
 SERVER_XML_FILE="$TOMCAT_PATH/conf/server.xml"
 CONTEXT_XML_FILE="$TOMCAT_PATH/conf/Catalina/$WEBSITE/ROOT.xml"
 
+ENGINE_XPATH="Service/Engine[@name=\"Catalina\"]"
+HOST_XPATH="Host[@name=\"$WEBSITE\"]"
+FULL_HOST_XPATH="$ENGINE_XPATH/$HOST_XPATH"
+
+if [ -n "$(read_xml_value "$SERVER_XML_FILE" "$FULL_HOST_XPATH" "appBase")" ]; then
+    "$(dirname "$0")/setup/delete-xml-node.py" "$SERVER_XML_FILE" "$ENGINE_XPATH" "$HOST_XPATH" || exit 1
+fi
+
 # Set up the config files and replace values as appropriate
 rm -rf "$TOMCAT_PATH/conf/Catalina/$WEBSITE"
 rm -rf "$TOMCAT_PATH/conf/Catalina/sitename.onevizion.com"
